requirements.txt
# Core backend
fastapi==0.111.0
uvicorn[standard]==0.22.0
python-multipart==0.0.7
pydantic==1.10.13
aiofiles==23.1.0

# Document parsing
python-docx==0.8.11
pdfminer.six==20221105

# Hugging Face & ML
transformers==4.44.2
sentence-transformers==3.0.1
torch>=2.0.0
numpy==1.26.4
scikit-learn==1.5.1

# Streamlit frontend
streamlit==1.38.0
requests==2.31.0

# Environment management
python-dotenv==1.0.0
